+++ src/lib.rs
@@ -21,5 +21,4 @@ pub use error::*;
 pub mod logging;
 pub use logging::*;
 
-pub mod hot_reload;
 pub mod plugin;
